Add docs for Torrent Checker
